New Art, School of Fish
Added new pixel art for the Lake background, new water assets, and the beginnings of the main character design.

School of fish code has been created and is fully functioning with a new fish type. May have to separate code that can be generalized into it's own class to be reused for other fish types.