New Fish Art, School of Fish Surface Avoidance, Background Fix
Added new artwork for fish (Minnows and Salmon). Minnows will be the smaller school of fish, and the Salmon will be the basic, larger fish.

Added some lines of code to allow schools of fish to avoid surfacing. Also made it a bit more performant, not calculating and weighing directions to maintain it's pull towards the school when it's going to abandon it all anytime it sees a food source.

Corrected an asset option in the background art files, causing certain zoom levels to have a thin line between the bottom of the "main" background and the portion that begins tiling.