More Fishables, Animations, Input System
Added a significant amount of art, including:
- Character basic animations
- Fishing rod animations
- New fishables art

Created new prefabs of different fishable types, working with behaviours already created to test the process of the plug-and-play ability of said behaviours.

Added the foundation of the shift over to Unity's Input System, allowing for easier handling and customization of player input.