Tutorial and Bait additions
Added a new tutorial system. This is similar to the tooltips system, but the game will fade slightly to black and slow down, giving the player a chance to read the message until they click close or wait a certain amount of time.

Also added the necessary prefabs and scriptables for bait to be made with the new fishables added.