Off Screen Optimization, Glow, Frame Rate
Added a script for all fishables that will disable and remove their behaviours from running when not needed. This is currently only activated by distance to the hook, and includes:
- The food search job
- School job
- Movement
- Hunger/Growth

Added "glow" sprites to make the player's visibility of fish in the water a bit easier as they get closer to them. Will likely either remove this feature or rework it into something else, as the fish aren't too hard to notice with a small amount of size increases, and the glow might be distracting if anything. May be useful for something in the future though.

Added a few lines of code to the Frame Rate Counter script which will allow assignment of how many frames to wait before updating the average frame rate since the previous update. These intervals should be easier to make sense of compared to updates every frame.