Min/Max Trajectories
Added UI for the minimum and maximum power trajectories.

These will display from the start of the casting process through to the cast, giving the player an idea of what angles provide which possibilities before they lock it in and begin choosing the power of the cast.

Still need to adjust camera size to better view the furthest distances.