Camera Update
Camera is no longer locked to the hook. It will freely move and adjust based on what's necessary to see at different points of time.

When idle, the camera has a default spot it'll rest at in front of the player, still allowing to zoom in/out.

When casting, the camera controls will be locked. When in the angling phase, it will zoom out and position to show the furthest possible cast you can accomplish with your equipped rod. Once that's set and you're choosing your power, it'll adjust it's zoom and position again to show the furthest possible cast you can accomplish with that specific angle.

As soon as you cast, the camera controls unlock and it adjusts back to how the player had it zoomed previously, and loosely but consistently follows the hook.

Both position and zoom lerp to a "desired" variable, giving a more smooth transition. However, I may still have to play with how to get the hook to look less jittery in camera, as any higher lerp speeds cause this issue.