Reeling Minigame
Added a basic form of the Stardew Valley type reeling minigame that initiates when you've hooked a fish.

- The goal is to get the green region to envelop the fish icon, which will reel the fish in while it's inside.

- The fish will have certain stats for the distance and frequency it changes it's icon's location in the minigame (still need to add those stats to each fish's prefab)

- Fishing rods have their own stats for reeling region size, some of the physics behind moving it, etc etc.

I'll soon want to add more depth to the minigame.

One will be adding a strain limit to the fishing line when you're moving the reel zone, resulting in less strain when it's in the correct position, and strain will go down over time when letting it loose.

This'd work in conjunction with another future thing to add by having fish attempting to escape, both in physically swimming away every so often for an amount of time, which also adds extra strain to the line. These will add a certain timing requirement for when you line up the green with the fish icon.