Line Stress
Added a stress mechanic to the reeling minigame, which takes into account the fish's base strength and the fishing line's strength to introduce difficulty.

If the player is pressing the reel button (to adjust the reel zone), a certain amount of stress will be introduced. Once the fish icon is in the reel zone, this won't apply.

When either not reeling or when the fish is within the reel zone, stress will decay.

The fill color of the minigame reflects the stress level of the fishing line.

Should the stress levels get to be more than the line's strength, the line snaps and returns the hook back to the player and destroys whatever was hooked (this will have to be reworked to just letting the fish go)