Hooked edibility and non-fish minigame
Added a variable for if a fishable swims or not, for use in the minigame which now checks if it should attempt to swim away or not.

Code now removes fishables from the food search Job upon being hooked. This alleviates potential frustration of hooking something and then having it eaten before you reel it in.