New Tiled Terrain
Implemented a new workflow for terrain, as well as prototype art for it. This workflow includes creating at tileset in Aseprite, using Tiled to create rules and the map, and importing it all to Unity with SuperTiled2Unity.

Previous scripts relied on a single PolygonCollider2D, which Tiled generates multiple of. As such, some changes were made to FishMovement, GroundMovement, and SpawnBed to work with an array. I also created a helper class Utilities, in this case for use of ClosestPointFromColliders(), which returns a ClosestPointInfo struct containing info on the position, distance, and collider for the closest point.

Additional work needs to be done to split some angled tiles into 2, to allow for one half to be drawn in the background. Normal mapping also needs to be done to pave the way for dynamic lighting.