Isometric Tile Z-Index fix/Unity Update
Updated Unity from a 2019 version to 2022.3.32f1 to attempt to fix a Z-Indexing issue with the isometric tileset that many other users seemed to have around 2019/2020. Unfortunately this wasn't the fix required, but I did find the solution (2D Renderer needed it's Transparency Sort Axis Z value to be Tilesize * -0.5 + 0.1). Fortunately, this means Unity is updated and most things still function.

This does require the removal of the Water2D Surface shader, as it relies on GrabPass which was a removed feature of shaders. However as this was an imported asset anyway, the plan was to replace this, and fiddling with custom shaders in Shader Graph is on the docket pretty soon. I also created a new Fishing Line material for fishing rod's LineRenderers to use.