Script Cleanup, FishScool/Shoal optimization and update
Began work on another pass of updating scripts to current practices. This included the Camera scripts, and all of the Fish Behaviour scripts besides Wander.

Fixed some behaviour in both FishSchoolBehaviour and Shoal. This both improved fish navigation and optimization, and now they have further debug features in the form of gizmos.