v0.16.1h - BackgroundParallax update
Improved the functionality of the BackgroundParallax script.

A new class has been created for ParallaxElement, which will do a better job of keeping track of each "layer" to be adjusted. This holds the transform, parallax strength, and the original position it starts from.

Now, BackgroundParallax holds just a list of these and performs the business logic, which is now under the AdjustPrallaxPositions method and called by FixedUpdate instead of Update.

Tooltips have also been added to each variable exposed to the inspector.

The script is now attached to the Fishing Level Main Camera GameObject instead of being attached to the object holding the background elements.