v0.17.1k - Removed IEdible, ISpawn reduced
Removed the IEdible script, and the Spawn() contract for ISpawn. All IEdible was managing was a Despawn method which was instead converted to using OnDestroyed for applicable objects. The Spawn method had no necessity to be called outside of their respective classes, so it was unnecessary to have it as a contract in an interface. All scripts referencing have bene updated to reflect these changes.