v0.16.1u - Updated Inventory UI scripts
Updated scripts in the UI/Inventory folder to match current practices.

Created the InactiveSingleton class in order to set static instance references and get dependencies on scripts that start off inactive, so that other scripts can reference them as needed.

Changed users of the BlockCastUI script over to DisableCastOnHover. Deleted BlockCastUI.

Added an action for ChangedEquippedRod for relevant scripts to subscribe to.

Defined calls to OnEnable and OnDisable in various scripts.

Deleted the IInventoryTab interface as its usage only created contracts that have been deprecated by using OnEnable and OnDisable.

Created the BaitUI abstract class, which different bits of UI can inherit from when they need to display different baits.

Created the AttachmentSlotButton class, which can be inherited from (as BaitAttachmentButton does), but handles most of the responsibilities needed for attachment buttons.