v0.16.1w - Updated RecordsMenu scripts
Updates scripts in the UI/Records folder to match current practices. This included an update to the RecordSaveData of PlayerData, which now generates a data point for each object in FishableScriptables of ItemLookupTable at the start of a new game. The records menu also now generates each listing in a similar fashion instead of having to manually update it with each fishable added/changed to the project.