v0.16.1y - Updated Title Menu scripts
Updated scripts in the UI/Title Menu folder to match current practices.

Removed TitleMenuManager and created additional scripts for each menu that will house its correlated functionality. The SwapActive method was moved over to Utilities.

Created prefabs and a folder containing them for the title screen's menus.

Made an InactiveSingletonCaller class, so that it can be attached to objects in scenes that may have need of setting static references of inactive singletons without the need of a UIManager.